CommandContext moved to core
